pyproject.toml
[project]
name = "gx-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.20",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.6",
    "langchain-tavily>=0.2.11",
    "langchain-text-splitters>=0.3.9",
    "langsmith>=0.4.15",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.0",
    "unstructured>=0.18.14",
    "uvicorn>=0.35.0",
]
